Refactoring; fuse SyntaxTestParser and SyntaxTester to SyntaxTest.
